+++ database/data/development-data/index.js
@@ -4,4 +4,4 @@ exports.groups = require("./groups.js");
 exports.userGroups = require("./userGroups.js");
 exports.events = require("./events.js");
 exports.eventUsers = require("./eventUsers.js");
-exports.eventComments = require("./eventComments.js");
+exports.eventComments = require("./comments.js");
